Move platform popup utils out from popup.
