+++ apps/browser/src/auth/services/extension-two-factor-auth-component.service.spec.ts
@@ -22,7 +22,7 @@ import { DuoLaunchAction } from "@bitwarden/auth/angular";
 import { TwoFactorProviderType } from "@bitwarden/common/auth/enums/two-factor-provider-type";
 
 import { BrowserApi } from "../../platform/browser/browser-api";
-import BrowserPopupUtils from "../../platform/popup/browser-popup-utils";
+import BrowserPopupUtils from "../../platform/browser/browser-popup-utils";
 import {
   AuthPopoutType,
   closeSsoAuthResultPopout,
