+++ apps/browser/src/auth/services/extension-two-factor-auth-component.service.ts
@@ -6,7 +6,7 @@ import {
 import { TwoFactorProviderType } from "@bitwarden/common/auth/enums/two-factor-provider-type";
 
 import { BrowserApi } from "../../platform/browser/browser-api";
-import BrowserPopupUtils from "../../platform/popup/browser-popup-utils";
+import BrowserPopupUtils from "../../platform/browser/browser-popup-utils";
 import {
   AuthPopoutType,
   closeSsoAuthResultPopout,
