+++ apps/browser/src/auth/services/extension-two-factor-auth-email-component.service.ts
@@ -7,7 +7,7 @@ import { ConfigService } from "@bitwarden/common/platform/abstractions/config/co
 import { DialogService } from "@bitwarden/components";
 
 import { openTwoFactorAuthEmailPopout } from "../../auth/popup/utils/auth-popout-window";
-import BrowserPopupUtils from "../../platform/popup/browser-popup-utils";
+import BrowserPopupUtils from "../../platform/browser/browser-popup-utils";
 
 // TODO: popup state persistence should eventually remove the need for this service
 export class ExtensionTwoFactorAuthEmailComponentService
