+++ apps/browser/src/platform/browser/abstractions/browser-popup-utils.abstractions.ts
