+++ apps/browser/src/vault/popup/settings/appearance-v2.component.ts
@@ -24,14 +24,12 @@ import {
   SelectModule,
 } from "@bitwarden/components";
 
+import { PopupWidthOption } from "../../../platform/browser/browser-popup-utils";
 import { PopOutComponent } from "../../../platform/popup/components/pop-out.component";
 import { PopupCompactModeService } from "../../../platform/popup/layout/popup-compact-mode.service";
 import { PopupHeaderComponent } from "../../../platform/popup/layout/popup-header.component";
 import { PopupPageComponent } from "../../../platform/popup/layout/popup-page.component";
-import {
-  PopupWidthOption,
-  PopupSizeService,
-} from "../../../platform/popup/layout/popup-size.service";
+import { PopupSizeService } from "../../../platform/popup/layout/popup-size.service";
 import { VaultPopupCopyButtonsService } from "../services/vault-popup-copy-buttons.service";
 
 @Component({
