+++ apps/browser/src/vault/popup/utils/vault-popout-window.ts
@@ -3,7 +3,7 @@
 import { CipherType } from "@bitwarden/common/vault/enums";
 
 import { BrowserApi } from "../../../platform/browser/browser-api";
-import BrowserPopupUtils from "../../../platform/popup/browser-popup-utils";
+import BrowserPopupUtils from "../../../platform/browser/browser-popup-utils";
 
 const VaultPopoutType = {
   viewVaultItem: "vault_viewVaultItem",
